refactor: move some class to base package folder
